Added tests for hash computation (for A and B file path hashes, at least). Fixed bug in the hashing algorithm.

+++ Nmpq/Parsing/HashType.cs
@@ -1,5 +1,5 @@
 namespace Nmpq.Parsing {
-	public enum HashType : ulong {
+	public enum HashType : uint {
 		TableOffset = 0,
 		FilePathA = 1,
 		FilePathB = 2,
